# docs/swagger.yaml
basePath: /api/
definitions:
...
  Asset:
    properties:
      logo_image_link:
        type: string
      logo_link:
        type: string
    type: object
...
  Controls:
    properties:
      enable_api:
        type: boolean
      enable_controls:
        type: boolean
      force_autoplay:
        type: boolean
      force_loop:
        type: boolean
      hide_title:
        type: boolean
    type: object
...
  GetMediaCaptionsData:
    properties:
      total:
        type: integer
      video_captions:
        items:
          $ref: '#/definitions/MediaCaption'
        type: array
    type: object
  GetMediaCaptionsResponse:
    properties:
      data:
        $ref: '#/definitions/GetMediaCaptionsData'
      status:
        type: string
    type: object
...
  GetMediaDetailResponse:
    properties:
      data:
        $ref: '#/definitions/Media'
      status:
        type: string
    type: object
...
  Media:
    properties:
      assets:
        $ref: '#/definitions/MediaAssets'
      captions:
        items:
          $ref: '#/definitions/MediaCaption'
        type: array
      chapters:
        items:
          $ref: '#/definitions/MediaChapter'
        type: array
      created_at:
        type: string
      description:
        type: string
      duration:
        type: number
      id:
        type: string
      is_mp4:
        type: boolean
      is_public:
        type: boolean
      metadata:
        items:
          $ref: '#/definitions/Metadata'
        type: array
      player_theme:
        $ref: '#/definitions/PlayerTheme'
      player_theme_id:
        type: string
      qualities:
        items:
          $ref: '#/definitions/QualityObject'
        type: array
      size:
        type: integer
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      view:
        type: integer
    type: object
  MediaAssets:
    properties:
      dash_iframe:
        type: string
      dash_player_url:
        type: string
      dash_url:
        type: string
      hls_iframe:
        type: string
      hls_player_url:
        type: string
      hls_url:
        type: string
      mp4_url:
        type: string
      source_url:
        type: string
      thumbnail_url:
        type: string
    type: object
  MediaCaption:
    properties:
      description:
        type: string
      is_default:
        type: boolean
      language:
        type: string
      status:
        type: string
      url:
        type: string
    type: object
  MediaChapter:
    properties:
      language:
        type: string
      url:
        type: string
    type: object
...
  Metadata:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  PlayerTheme:
    properties:
      asset:
        $ref: '#/definitions/Asset'
      controls:
        $ref: '#/definitions/Controls'
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      theme:
        $ref: '#/definitions/Theme'
      user_id:
        type: string
    type: object
...
  QualityObject:
    properties:
      audio_bitrate:
        type: integer
      audio_codec:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      video_bitrate:
        type: integer
      video_codec:
        type: string
    type: object
...
  ResponseError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  ResponseSuccess:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  Theme:
    properties:
      control_bar_background_color:
        type: string
      control_bar_height:
        type: string
      main_color:
        type: string
      menu_background_color:
        type: string
      menu_item_background_hover:
        type: string
      progress_bar_circle_size:
        type: string
      progress_bar_height:
        type: string
      text_color:
        type: string
      text_track_background:
        type: string
      text_track_color:
        type: string
    type: object
...
  UpdateMediaInfoRequest:
    properties:
      description:
        description: Description of the media
        type: string
      is_public:
        description: Media's publish status
        type: boolean
      metadata:
        description: Media's metadata
        items:
          $ref: '#/definitions/Metadata'
        type: array
      player_id:
        description: Media player 's id
        type: string
      tags:
        description: Media's tags
        items:
          type: string
        type: array
      title:
        description: Title of the media
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aioz Stream Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aioz Stream API
  version: "1.0"
paths:
...
  /videos/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a media by media ID.
      operationId: DELETE_media-mediaId
      parameters:
        - description: Media ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Delete media successfully.
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: Available requests left for current window
              type: integer
            X-RateLimit-Retry-After:
              description: Seconds until rate limit window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: Available requests left for current window
              type: integer
            X-RateLimit-Retry-After:
              description: Seconds until rate limit window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: Available requests left for current window
              type: integer
            X-RateLimit-Retry-After:
              description: Seconds until rate limit window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: Available requests left for current window
              type: integer
            X-RateLimit-Retry-After:
              description: Seconds until rate limit window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: Available requests left for current window
              type: integer
            X-RateLimit-Retry-After:
              description: Seconds until rate limit window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - Bearer: [ ]
        - BasicAuth: [ ]
      summary: Delete media
      tags:
        - media
      x-client-action: delete
    get:
      description: Retrieve the media details by media id.
      operationId: GET-media-status
      parameters:
        - description: media's id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/GetMediaDetailResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - BasicAuth: [ ]
        - Bearer: [ ]
      summary: get media detail
      tags:
        - media
      x-client-action: getDetail
    patch:
      consumes:
        - application/json
      operationId: PATCH_media
      parameters:
        - description: media's id
          in: path
          name: id
          required: true
          type: string
        - description: input
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/UpdateMediaInfoRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - BasicAuth: [ ]
        - Bearer: [ ]
      summary: update media info
      tags:
        - media
      x-client-action: update
  /videos/{id}/captions:
    get:
      consumes:
        - application/json
      description: Retrieves a list of media captions for the specified media.
      operationId: GET_media-mediaId-caption-language
      parameters:
        - description: Media ID
          in: path
          name: id
          required: true
          type: string
        - default: 0
          description: offset, allowed values greater than or equal to 0. Default(0)
          in: query
          minimum: 0
          name: offset
          type: integer
        - default: 25
          description: results per page. Allowed values 1-100, default is 25
          in: query
          maximum: 100
          minimum: 1
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/GetMediaCaptionsResponse'
        "400":
          description: Bad Request
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "403":
          description: Forbidden
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "404":
          description: Not Found
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
        "500":
          description: Internal Server Error
          headers:
            X-RateLimit-Limit:
              description: The request limit per minute
              type: integer
            X-RateLimit-Remaining:
              description: The number of available requests left for the current time
                window
              type: integer
            X-RateLimit-Retry-After:
              description: The number of seconds left until the current rate limit
                window resets
              type: integer
          schema:
            $ref: '#/definitions/ResponseError'
      security:
        - BasicAuth: [ ]
        - Bearer: [ ]
      summary: Get media captions
      tags:
        - media
      x-client-action: getCaptions
      x-client-paginated: true
      x-group-parameters: true
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"